requirements2.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
cryptography==46.0.4
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.128.6
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
intel-openmp==2021.4.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-mcp-adapters==0.2.0
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.15
langsmith==0.7.1
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mcp==1.26.0
mkl==2021.4.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
openai==2.18.0
opencv-python==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
parso==0.8.6
pillow==11.3.0
platformdirs==4.5.1
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.17.0
sentence-transformers==2.6.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tbb==2021.13.1
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1+cpu
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.41.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.6.0
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
